Refactor: Organize data sources and update pupil detail screen
This commit refactors the data source organization by moving files into `data/sources/local` and `data/sources/network` packages.

It also includes the following updates to the pupil detail screen:
- Improved handling of pupil not found scenarios.
- Enhanced error handling with more specific messages.
- Added Timber logging for better debugging.
- Minor UI adjustments for a cleaner presentation.

Additionally, ApplicationTest.java was removed.

+++ android-tech-test-kotlin/app/src/main/java/com/bridge/androidtechnicaltest/feature/pupil/components/PupilRecyclerAdapter.kt
@@ -40,7 +40,7 @@ class PupilRecyclerAdapter(
                 val position = bindingAdapterPosition
                 if (position != RecyclerView.NO_POSITION) {
                     val pupil = getItem(position)
-                    pupilAction.navigateToPupilDetail(pupil.pupilId.toInt())
+                    pupilAction.navigateToPupilDetail(pupil.pupilId)
                 }
             }
         }
